% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.r
\name{same_rd_name.default}
\alias{same_rd_name.default}
\alias{same_rd_name.data.frame}
\title{Test default}
\usage{
\method{same_rd_name}{default}(x)

\method{same_rd_name}{data.frame}(x)
}
\arguments{
\item{x}{A parameter}

\item{x}{A parameter}
}
\description{
Test default

Test data frame
}
